Remove the OAS contract monitoring related code from `tap/api.go`
